+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fix a bug: enable admins to be invited to collaborate on projects
 - Fix a bug: hide some links to create an account if public registrations is disabled
 - Fix a bug: unable to save user validation if admin click save user profile button after switch user valitation
+- Fix a bug: if multi VAT no value is filled in, the general rate can't apply
 - Fix a security issue: updated rack to 2.2.3.1 to fix [CVE-2022-30123](https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2022-30123) and [CVE-2022-30122](https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2022-30122)